+++ src/main/java/kdkd/youre/backend/domain/category/domain/Category.java
@@ -1,12 +1,10 @@
-package kdkd.youre.backend.domain.category.repository;
+package kdkd.youre.backend.domain.category.domain;
 
-import kdkd.youre.backend.domain.member.repository.Member;
+import kdkd.youre.backend.domain.member.domain.Member;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
 
 import javax.persistence.*;
