+++ src/main/java/kdkd/youre/backend/domain/member/domain/Member.java
@@ -1,4 +1,4 @@
-package kdkd.youre.backend.domain.member.repository;
+package kdkd.youre.backend.domain.member.domain;
 
 import lombok.*;
 
@@ -13,9 +13,14 @@ public class Member {
     @Column(name = "member_id")
     private Long id;
 
-    private String nickname; // 회원 이름
+    // 회원 기본 정보
+    private String nickname; // 회원 닉네임
     private String email; // 회원 이메일
+
+    // 소셜 로그인 정보
     private String oauthProvider; // oAuth Provider (google)
+    private String loginId; // 로컬 로그인용. username 역할
+    private String password; // 로컬 로그인용
 
     @Builder
     public Member(
