+++ src/main/java/kdkd/youre/backend/domain/member/domain/repository/MemberRepository.java
@@ -1,0 +1,9 @@
+package kdkd.youre.backend.domain.member.domain.repository;
+
+import kdkd.youre.backend.domain.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Member findByLoginId(String loginId);
+}
