--- src/main/java/kdkd/youre/backend/domain/member/repository/MemberRepository.java
@@ -1,7 +1,0 @@
-package kdkd.youre.backend.domain.member.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-}
